refactor(environment): change directory of environment and utility functions
