--- src/environment/utils/types/validate-options.type.ts
@@ -1,7 +1,0 @@
-import { BaseEnvironment } from '../../environment.type';
-import { CoreEnvironmentModuleOptions } from '../../types';
-
-export type ValidateOptions<TEnvironment extends BaseEnvironment> = Pick<
-  CoreEnvironmentModuleOptions<TEnvironment>,
-  'validationSchema' | 'environmentClass'
-> & { configuration: Record<string, unknown> };
