+++ src/lib/environment/types/core-environment-module-options.type.ts
@@ -3,7 +3,7 @@ import { BaseEnvironment } from '../environment.type';
 import { ValidationSchema } from './validation-schema.type';
 
 export type CoreEnvironmentModuleOptions<TEnvironment extends BaseEnvironment> =
-  {
+  Readonly<{
     environmentClass: ClassConstructor<TEnvironment>;
     validationSchema: ValidationSchema<TEnvironment>;
-  };
+  }>;
