+++ src/util/validation/types/validate-options.type.ts
@@ -1,0 +1,13 @@
+import {
+  BaseEnvironment,
+  CoreEnvironmentModuleOptions,
+} from '@lib/environment';
+
+export type ValidateOptions<TEnvironment extends BaseEnvironment> = Readonly<
+  {
+    configuration: Record<string, unknown>;
+  } & Pick<
+    CoreEnvironmentModuleOptions<TEnvironment>,
+    'validationSchema' | 'environmentClass'
+  >
+>;
